requirements.txt
telethon==1.32.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
python-magic==0.4.27
loguru==0.7.2 